+++ CHANGELOG.md
@@ -47,6 +47,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
   * Move `rpc/ethereum/backend` -> `rpc/backend`
   * Move `rpc/ethereum/namespaces` -> `rpc/namespaces/ethereum`
 
+### Improvements
+
+* (ci, evm) [tharsis#1063](https://github.com/tharsis/ethermint/pull/1063) Run simulations on CI.
+
 ### Bug Fixes
 
 * (rpc) [tharsis#1059](https://github.com/tharsis/ethermint/pull/1059) Remove unnecessary event filtering logic on the `eth_baseFee` JSON-RPC endpoint.